Mover verification controllers and view
