+++ scout/server/blueprints/variant/controllers.py
@@ -51,7 +51,6 @@ def variant(store, institute_id, case_name, variant_id=None, variant_obj=None, a
         }
 
     """
-    variant_type = variant_type or 'snv'
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
     # If the variant is already collected we skip this part
     if not variant_obj:
@@ -61,6 +60,7 @@ def variant(store, institute_id, case_name, variant_id=None, variant_obj=None, a
     if variant_obj is None:
         return None
 
+    variant_type = variant_type or variant_obj.get('category', 'snv')
     variant_id = variant_obj['variant_id']
 
     genome_build = case_obj.get('genome_build', '37')
