+++ tests/server/blueprints/variants/test_variants_controllers.py
@@ -1,45 +1,5 @@
-from scout.server.blueprints.variants.controllers import (variant_verification, 
-variants_export_header, variant_export_lines)
-
-def url_for(param, institute_id, case_name, variant_id):
-    pass
-
-
-def test_sanger_mail_sent(mock_mail, real_variant_database, institute_obj, case_obj, user_obj,
-                          mock_sender, mock_comment):
-    adapter = real_variant_database
-    ## GIVEN we have a variant the we want to order sanger for
-    variant_obj = adapter.variant_collection.find_one()
-    variant_obj['hgnc_symbols'] = ''
-    variant_obj['panels'] = ''
-
-    ## WHEN calling variant_verification method with order==True
-    variant_verification(adapter, mock_mail, institute_obj, case_obj, user_obj, variant_obj, 
-                        mock_sender, 'complete_variant_url', True, mock_comment, 
-                        url_builder=url_for)
-
-    ## THEN the supplied mail objects send method should have been called
-    assert mock_mail._send_was_called
-    ## THEN the supplied mail objects send method should have received a message object
-    assert mock_mail._message
-
-def test_cancel_sanger_mail_sent(mock_mail, real_variant_database, institute_obj, case_obj, user_obj,
-                                  mock_sender, mock_comment):
-    adapter = real_variant_database
-    ## GIVEN we have a variant the we want to order sanger for
-    variant_obj = adapter.variant_collection.find_one()
-    variant_obj['hgnc_symbols'] = ''
-    variant_obj['panels'] = ''
-
-    ## WHEN calling variant_verification method with order==False
-    variant_verification(adapter, mock_mail, institute_obj, case_obj, user_obj, variant_obj, 
-                         mock_sender, 'complete_variant_url', False, mock_comment, 
-                         url_builder=url_for)
-
-    ## THEN the supplied mail objects send method should have been called
-    assert mock_mail._send_was_called
-    ## THEN the supplied mail objects send method should have received a message object
-    assert mock_mail._message
+from scout.server.blueprints.variants.controllers import (variants_export_header, 
+                                                          variant_export_lines)
 
 def test_variant_csv_export(real_variant_database, case_obj):
     adapter = real_variant_database
